DECRBY, HINCRBY, and INCRBY should all accept Long values
The documentation for all three of these commands states:

  "[The operation] is limited to 64 bit signed integers."

 * http://redis.io/commands/hincrby
 * http://redis.io/commands/decrby
 * http://redis.io/commands/incrby

+++ src/main/scala/com/redis/HashOperations.scala
@@ -25,7 +25,7 @@ trait HashOperations { self: Redis =>
       }
     }
 
-  def hincrby(key: Any, field: Any, value: Int)(implicit format: Format): Option[Long] =
+  def hincrby(key: Any, field: Any, value: Long)(implicit format: Format): Option[Long] =
     send("HINCRBY", List(key, field, value))(asLong)
 
   def hincrbyfloat(key: Any, field: Any, value: Float)(implicit format: Format): Option[Float] =
